requirements.txt
google-generativeai
opencv-python
gtts
gradio
playsound==1.2.2